added limited support for fullsize flexboard
Added limited support for the fullsize Flexboard configuration- many
keys are not mapped to anything useful in the keymap, but the current
configuration can be used for bringup and testing.

Also, enabled fake "USB boot protocol support" for the cut
configuration. The ZMK firmware does not seem to support the boot
protocol, but if we lie to the OS and say that support exists than Linux
will enable wakeup support for the keyboard by default (allowing us to
wake the system from sleep)

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>